converting encoders to half precision

+++ models/BrownianBridge/LatentBrownianBridgeModel.py
@@ -25,7 +25,7 @@ class LatentBrownianBridgeModel(BrownianBridgeModel):
         self.radar_encoder = VoxelNet()
         encoder_checkpoint = torch.load(model_config.RADAR_ENCODER.checkpoint_path)
         self.radar_encoder.load_state_dict(encoder_checkpoint['radar_encoder_state_dict'])
-
+        self.radar_encoder = self.radar_encoder.half()
         self.lidar_encoder = Lidar_VAE(model_config.LIDAR_ENCODER.config_path, 
                                        model_config.LIDAR_ENCODER.checkpoint_path,
                                        device)
